+++ services/gateway/cmd/server.go
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2024
+ * (c) Copyright Ascensio System SIA 2026
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -52,12 +52,14 @@ func Server() *cli.Command {
 				pkg.WithModules(
 					controller.NewAuthController, controller.NewEditorController,
 					controller.NewFileController, controller.NewShareController,
+					controller.NewDesktopController,
 					chttp.NewService, web.NewServer,
 					shared.BuildNewIntegrationCredentialsConfig(CONFIG_PATH),
 					shared.NewBoxAPIClient, shared.BuildNewOnlyofficeConfig(CONFIG_PATH),
 					crypto.NewStateGenerator,
 					middleware.NewSessionStore,
 					middleware.NewSessionMiddleware,
+					middleware.NewDesktopMiddleware,
 					format.NewMapFormatManager,
 				),
 			).Bootstrap()