+++ services/shared/response/config.go
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2024
+ * (c) Copyright Ascensio System SIA 2026
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -68,6 +68,7 @@ type EditorConfig struct {
 	CallbackURL   string        `json:"callbackUrl"`
 	Customization Customization `json:"customization"`
 	Lang          string        `json:"lang,omitempty"`
+	Mode          string        `json:"mode,omitempty"`
 }
 
 type User struct {